LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/user-pass.https_cross-origin=true-expected.txt


FAIL test www-authenticate basic does not forward credentials to cross-origin pages. assert_equals: expected (undefined) undefined but got (string) "user"
